Write bluetooth connection to device for lamarzocco
