+++ homeassistant/components/lamarzocco/entity.py
@@ -6,8 +6,12 @@ from dataclasses import dataclass
 from pylamarzocco.const import FirmwareType
 from pylamarzocco.lm_machine import LaMarzoccoMachine
 
-from homeassistant.const import CONF_ADDRESS
-from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC, DeviceInfo
+from homeassistant.const import CONF_ADDRESS, CONF_MAC
+from homeassistant.helpers.device_registry import (
+    CONNECTION_BLUETOOTH,
+    CONNECTION_NETWORK_MAC,
+    DeviceInfo,
+)
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
@@ -48,17 +52,17 @@ class LaMarzoccoBaseEntity(
             serial_number=device.serial_number,
             sw_version=device.firmware[FirmwareType.MACHINE].current_version,
         )
+        connections: set[tuple[str, str]] = set()
         if coordinator.config_entry.data.get(CONF_ADDRESS):
-            self._attr_device_info.update(
-                DeviceInfo(
-                    connections={
-                        (
-                            CONNECTION_NETWORK_MAC,
-                            coordinator.config_entry.data[CONF_ADDRESS],
-                        )
-                    }
-                )
+            connections.add(
+                (CONNECTION_NETWORK_MAC, coordinator.config_entry.data[CONF_ADDRESS])
             )
+        if coordinator.config_entry.data.get(CONF_MAC):
+            connections.add(
+                (CONNECTION_BLUETOOTH, coordinator.config_entry.data[CONF_MAC])
+            )
+        if connections:
+            self._attr_device_info.update(DeviceInfo(connections=connections))
 
 
 class LaMarzoccoEntity(LaMarzoccoBaseEntity):
